Refactor abstract_main
